move stackdriver to it's own package
