+++ stackdriver/trace_proto.go
